environment_requirements.txt
# Name                    Version                   Build  Channel
boto3                     1.29.3                   pypi_0    pypi
botocore                  1.32.3                   pypi_0    pypi
bzip2                     1.0.8                h620ffc9_4  
ca-certificates           2023.08.22           hca03da5_0  
distro                    1.8.0                    pypi_0    pypi
jmespath                  1.0.1                    pypi_0    pypi
jpype1                    1.4.1                    pypi_0    pypi
libffi                    3.4.4                hca03da5_0  
ncurses                   6.4                  h313beb8_0  
numpy                     1.26.1                   pypi_0    pypi
openssl                   3.0.12               h1a28f6b_0  
packaging                 23.2                     pypi_0    pypi
pandas                    2.1.3                    pypi_0    pypi
pip                       23.3            py311hca03da5_0  
psycopg2-binary           2.9.9                    pypi_0    pypi
python                    3.11.5               hb885b13_0  
python-dateutil           2.8.2                    pypi_0    pypi
pytz                      2023.3.post1             pypi_0    pypi
pyyaml                    6.0.1                    pypi_0    pypi
readline                  8.2                  h1a28f6b_0  
s3transfer                0.7.0                    pypi_0    pypi
setuptools                68.0.0          py311hca03da5_0  
six                       1.16.0                   pypi_0    pypi
sqlalchemy                2.0.23                   pypi_0    pypi
sqlite                    3.41.2               h80987f9_0  
tabula-py                 2.8.2                    pypi_0    pypi
tk                        8.6.12               hb8d0fd4_0  
typing-extensions         4.8.0                    pypi_0    pypi
tzdata                    2023.3                   pypi_0    pypi
urllib3                   2.0.7                    pypi_0    pypi
wheel                     0.41.2          py311hca03da5_0  
xz                        5.4.2                h80987f9_0  
zlib                      1.2.13               h5a0b063_0  